# src/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "version": "1.0.0",
        "title": "Task management",
        "description": "An API to manage my personal tasks.",
        "contact": {
            "name": "Saul Dominguez",
            "url": "https://sauldom.dev/",
            "email": "saulydominguez@gmail.com"
        }
    },
    "paths": {
        "/auth/sign-up": {
            "post": {
                "description": "Sign up a new user",
                "tags": ["Auth"],
                "requestBody": {
                    "description": "The request body should contain an email and password for the new user",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SignUpRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User was signed up successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "default": "ok"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/auth/login": {
            "post": {
                "description": "Log in an existing user",
                "tags": ["Auth"],
                "requestBody": {
                    "description": "The request body should contain an email and password for the user",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User was logged in successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "default": "ok"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        {
            "apiKey": []
        }
    ],
    "components": {
        "securitySchemes": {
            "apiKey": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        },
        "schemas": {
...
            "LoginRequest": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "required": ["email", "password"]
            },
            "SignUpRequest": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "required": ["email", "password"]
            },
...
        }
    }
}